requirements.txt
numpy==1.23.5
pandas==2.0.0
scikit-learn==1.2.2
tensorflow==2.12.0
opencv-python-headless==4.7.0.72
scikit-image==0.20.0
imbalanced-learn==0.10.1
matplotlib==3.7.1
seaborn==0.12.2
gradio==3.33.1
joblib==1.2.0
tqdm==4.65.0
plotly==5.18.0
pillow==9.5.0
